app/requirements.txt
aioredis==2.0.1
django==4.2.4
psycopg2==2.9.7
django-redis==5.3.0
django-extensions==3.2.3
django-filter==23.3
django-mptt==0.14.0
django-mptt-admin==2.5.1
graylog-json-formatter
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
django-csp==3.7
coreapi==2.3.3
drf-yasg==1.21.7
ipython
requests
progress